dozens of fixes: floats, while

+++ Assignment2/FedirkoMykhnevych.A2/Expressions-Main/src/FedirkoMykhnevych/A2/OFPType.java
@@ -65,7 +65,7 @@ public class OFPType {
 		case "char":
 			return "char";
 		case "float":
-			return "float";
+			return "double";
 		case "bool":
 			return "boolean";
 		case "void":
